+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - fix: iOS13-Swift build (#1522)
+- fix: Check task support on setState: (#1523)
 - fix: Create span for loadView (#1495)
 - feat: Add flag to control network requests breadcrumbs (#1505)
 - feat: Support for ignored signals with SIGN_IGN (#1489)